[#61]Completed Aop and Encryption and Testing Left to complete.
