--- rupeshsunuwar/spingboot/MultipleDatabase/src/main/java/com/multipledatabase/MultipleDatabaseApplication.java
@@ -1,13 +1,0 @@
-package com.multipledatabase;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MultipleDatabaseApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(MultipleDatabaseApplication.class, args);
-    }
-
-}
