+++ rupeshsunuwar/springboot/MultipleDatabase/src/main/java/com/multipledatabase/inventory_db/entity/Product.java
@@ -1,0 +1,34 @@
+package com.multipledatabase.inventory_db.entity;
+
+
+import com.multipledatabase.inventory_db.enums.ProductEnum;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.CreationTimestamp;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+@Entity
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Product {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private int id;
+    private String name;
+    private String code;
+    @Enumerated(EnumType.STRING)
+    private ProductEnum status;
+    private double qty;
+    private double price;
+    @CreationTimestamp
+    private LocalDateTime created;
+
+
+
+
+}
