+++ rupeshsunuwar/springboot/product/src/main/java/com/restapi/product/ProductApplication.java
@@ -1,0 +1,15 @@
+package com.restapi.product;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cache.annotation.EnableCaching;
+
+@SpringBootApplication
+@EnableCaching
+public class ProductApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ProductApplication.class, args);
+    }
+
+}
